+++ include/matrix.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   matrix.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mkhan <mkhan@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 10:38:18 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/11/24 14:52:01 by mkhan            ###   ########.fr       */
+/*   Updated: 2022/11/27 13:56:32 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 # include "miniRT.h"
 
-typedef float			t_mat4[4][4];
-typedef float			t_mat3[3][3];
-typedef float			t_mat2[2][2];
+typedef double			t_mat4[4][4];
+typedef double			t_mat3[3][3];
+typedef double			t_mat2[2][2];
 
 typedef struct s_scene	t_scene;
 
@@ -26,11 +26,11 @@ void	mat_multiply(t_mat4 *res, const t_mat4 *m1, const t_mat4 *m2);
 void	mat_vec_multiply(t_vector *res, const t_mat4 *mat,
 			const t_vector *vec);
 void	transpose_matrix(t_mat4 *mat);
-void	translate_matrix(t_mat4 *mat, float x, float y, float z);
-void	scaling_matrix(t_mat4 *mat, float x, float y, float z);
+void	translate_matrix(t_mat4 *mat, double x, double y, double z);
+void	scaling_matrix(t_mat4 *mat, double x, double y, double z);
 
 // matrix_inverse.c
-float	det3(const t_mat3 *mat3);
+double	det3(const t_mat3 *mat3);
 void	submat4(t_mat3 *res, const t_mat4 *mat, int row, int col);
 void	mat_inverse(t_mat4 *res, const t_mat4 *mat);
