+++ include/vector.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 14:33:33 by hsarhan           #+#    #+#             */
-/*   Updated: 2022/11/22 16:04:15 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/11/27 13:56:32 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,22 +21,22 @@
 typedef struct s_vector	t_vector;
 struct s_vector
 {
-	float	x;
-	float	y;
-	float	z;
-	float	w;
+	double	x;
+	double	y;
+	double	z;
+	double	w;
 };
 
 // vector_arithmetic.c
 void	add_vec(t_vector *res, const t_vector *v1, const t_vector *v2);
 void	sub_vec(t_vector *res, const t_vector *v1, const t_vector *v2);
 void	negate_vec(t_vector *res, const t_vector *v);
-void	scale_vec(t_vector *res, const t_vector *v, float scale);
+void	scale_vec(t_vector *res, const t_vector *v, double scale);
 
 // vector_operations.c
-float	vec_magnitude(const t_vector *vec);
+double	vec_magnitude(const t_vector *vec);
 void	normalize_vec(t_vector *vec);
-float	dot_product(const t_vector *v1, const t_vector *v2);
+double	dot_product(const t_vector *v1, const t_vector *v2);
 void	cross_product(t_vector *res, const t_vector *v1, const t_vector *v2);
 
 #endif
