+++ src/math/matrix_transformations.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 15:41:22 by mkhan             #+#    #+#             */
-/*   Updated: 2022/11/27 13:46:07 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/11/27 13:56:32 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
  * @param z The value along the z-axis to be translated.
  */
 
-void	translate_matrix(t_mat4 *mat, float x, float y, float z)
+void	translate_matrix(t_mat4 *mat, double x, double y, double z)
 {
 	ft_bzero(mat, sizeof(t_mat4));
 	(*mat)[0][0] = 1;
@@ -39,7 +39,7 @@ void	translate_matrix(t_mat4 *mat, float x, float y, float z)
  * @param y The value along the y-axis to be scaled.
  * @param z The value along the z-axis to be scaled.
  */
-void	scaling_matrix(t_mat4 *mat, float x, float y, float z)
+void	scaling_matrix(t_mat4 *mat, double x, double y, double z)
 {
 	ft_bzero(mat, sizeof(t_mat4));
 	(*mat)[0][0] = x;
@@ -53,7 +53,7 @@ void	scaling_matrix(t_mat4 *mat, float x, float y, float z)
  * @param mat The matrix to be initialized
  * @param r angle in radians
  */
-void	rotation_matrix_x(t_mat4 *mat, float r)
+void	rotation_matrix_x(t_mat4 *mat, double r)
 {
 	ft_bzero(mat, sizeof(t_mat4));
 	(*mat)[0][0] = 1;
@@ -69,7 +69,7 @@ void	rotation_matrix_x(t_mat4 *mat, float r)
  * @param mat The matrix to be initialized
  * @param r angle in radians
  */
-void	rotation_matrix_y(t_mat4 *mat, float r)
+void	rotation_matrix_y(t_mat4 *mat, double r)
 {
 	ft_bzero(mat, sizeof(t_mat4));
 	(*mat)[0][0] = cos(r);
@@ -85,7 +85,7 @@ void	rotation_matrix_y(t_mat4 *mat, float r)
  * @param mat The matrix to be initialized
  * @param r angle in radians
  */
-void	rotation_matrix_z(t_mat4 *mat, float r)
+void	rotation_matrix_z(t_mat4 *mat, double r)
 {
 	ft_bzero(mat, sizeof(t_mat4));
 	(*mat)[0][0] = cos(r);
@@ -99,9 +99,9 @@ void	rotation_matrix_z(t_mat4 *mat, float r)
 /**
  * @brief Convert Radians to Degrees
  * @param r The radian value
- * @return float The converted value in degrees
+ * @return double The converted value in degrees
  */
-float	rad_to_deg(float r)
+double	rad_to_deg(double r)
 {
 	return(r * (180 / M_PI));
 }
@@ -109,9 +109,9 @@ float	rad_to_deg(float r)
 /**
  * @brief Convert Degrees to Radians
  * @param r The degree value
- * @return float The converted value in radians
+ * @return double The converted value in radians
  */
-float	deg_to_rad(float r)
+double	deg_to_rad(double r)
 {
 	return(r * (M_PI / 180));
 }
